Global: - added some texture files.
Camera.h and .cpp
- decreased default scale by 10 times;
- changed move() function to work with "in window" coordinates and current scale, for smoother camera movement;
- removed commented code.
main.cpp
- changed shapesize of created planets, previously diameter was used instead of radius;
- removed commented code of another variant of solar system initialization;
- changed pointer position variables from world coordinates Vector2f to window coordinates Vector2i;
- changed default view to center on Sun;
- translated some comments;
- added "if" statement for focusing on planetary systems;
- changed window views and draws order for both simulation variants to draw info first, for correct camera movement with mouse (previously info have benn drawing last, using default window view, that caused incorrect world coorinates for pointer on next input loop).

planet.h
- replaced separate X and Y coordinate members from class with Vectorld2d coordinate member;
- added global_coordinates member;
- added bounding_volume member for mouse pointer collision and future detection collision and possible draw selection;
- added ounds member for bounding_volume visualisation;
- rearranged class functions by groups for more readability;
- added setter for mass for future usage;
- added getters and setters for global coordinates both separate X and Y and Vectorld2d;
- added getter and setter for Vectorld2d coordinates;
- removed unneded draw function overload;

planet.cpp
- removed commented code from constructor, added nesessary code for Vectorld2d coords and global coords;
- rearranged function definitions to corelate with header file;
- added definitions for added functions;
- changed all functions to work with Vectorld2d coordinate members;
- replaced[] with .x and .y for better readability;
- added update function for bounding volume;
- updated draw function to work with global coordinates  and draw bounding volume shape for visualisation.

PlanetSystem.h
- replaced separate X and Y coordinate members of class with unified Vectorld2d one;
- added global_coord member;
- rearranged functions for better readability;
- added getters and setters for global_coords both for separate X and Y and unified one;
- removed add_planet() as its not needed for now;
- changed return type of baricenter() from std::vector to Vectorld2d;
- removed draw() overload.

PlanetSystem.cpp
- rearranged function definitions to corelate with header file for better readabilty;
- changed constructor to work with added members;
- changed funtions to work with new baricenter() return data type;
- added definitions of added functions;
- changed simulate() to take in account baricenter movement and to change global coordinates of planets according to global coordinate of planetary system and to update bounduing_volume.

SolarSystem.h
- replaced separate X and Y coordinate members with unified Vectorld2d one;
- rearranged functions by groups for better readability;
- added/changed getters and setter to work with unified coordinate;
- removed add_system() as its no needed for now;
- changed baricenter() to return Vectorld.2d.

SolarSystem.cpp
- rearranged definitions to corelate with header file;
- changed constructor and other functions to work with unified Vectorld2d coordinate;
- added getter and setter for unified coordinate;
- changed constructor to assign correct global coordinates to all planetary systems and planets and update their bounding volumes;
- updated simulate() to change planetary system's global coordinates and update their bounding_volumes after all movement is calculated;
- added interation_list() function euivalent to same functions in planet and PlanetSystem classes;

Vectorld2d.h and .cpp
- changed return type of overloaded + and * operators from reference to object;.

Volume.h and .cpp
- added coments;
- added setter for coordinates;
- corrected isPointInVolume() function: changed comparison operators to opposite as they were wrong.

+++ Camera.h
@@ -30,7 +30,7 @@ public:
 	//resizes after change of size of window
 	void resize(int newWidth, int newHeight);
 	//camera move
-	void move(sf::Vector2f oldPosition, sf::Vector2f newPosition);
+	void move(sf::Vector2i oldPosition, sf::Vector2i newPosition);
 	//resets camera to default 
 	void reset();
 	//changes camera scale
